refactor: 마이페이지 기능 파일 구조 더 직관적으로 수정
